# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TTaxi Platform API
  version: 1.0.0
  description: |
    Thailand Travel Platform — Airport Transfer & future travel services.
    Aligns with Database Design v1.1 and API Contract v1.0.
    Base path `/api/v1`. JWT Bearer authentication.
  contact:
    name: TTaxi API Support
    email: support@ttaxi.com
...
servers:
  - url: https://api.ttaxi.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /bookings/{bookingNumber}/notifications:
    get:
      tags: [Booking]
      summary: List booking-scoped notifications
      description: >
        Authenticated customer owner or guest via X-Guest-Access-Token header.
        Guest access returns only notifications for this booking.
      operationId: listBookingNotifications
      security:
        - bearerAuth: []
        - guestReviewAccessHeader: []
      parameters:
        - $ref: '#/components/parameters/BookingNumber'
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: notificationType
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNotificationList'
        '403':
          $ref: '#/components/responses/BookingNotAccessible'
  /customer/notifications:
    get:
      tags: [Customer]
      summary: List authenticated customer notifications
      operationId: listCustomerNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: notificationType
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNotificationList'
...
  /admin/notifications:
    get:
      tags: [Admin]
      summary: List admin notifications
      description: Each admin sees only notifications addressed to their user account.
      operationId: listAdminNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: notificationType
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNotificationList'
...
  /driver/notifications:
    get:
      tags: [Driver]
      summary: List driver notifications
      operationId: listDriverNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: notificationType
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNotificationList'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. Obtain via /auth/login or /auth/register.
    guestBookingAccess:
      type: apiKey
      in: query
      name: guestAccessToken
      description: Guest booking access token with email query param.
    guestReviewAccessHeader:
      type: apiKey
      in: header
      name: X-Guest-Access-Token
      description: Guest booking access token for review lookup. Never send in URL query parameters.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    BookingNumber:
      name: bookingNumber
      in: path
      required: true
      schema:
        type: string
        pattern: '^TX[0-9]{12}$'
      example: TX202607010001
...
  responses:
...
    BookingNotAccessible:
      description: Booking access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            denied:
              value:
                success: false
                error_code: FORBIDDEN
                message: Booking not accessible
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - INVALID_STATUS_TRANSITION
        - DUPLICATE_BOOKING
        - FLIGHT_NOT_FOUND
        - GOOGLE_PLACE_NOT_FOUND
        - INTERNAL_SERVER_ERROR
        - AUTH_REQUIRED
        - AUTH_INVALID
        - VEHICLE_NOT_SELECTABLE
        - DRIVER_NOT_AVAILABLE
        - BOOKING_NOT_ACCESSIBLE
        - RATE_LIMIT
        - EXTERNAL_API_ERROR
        - INVALID_FLIGHT_NUMBER
        - INVALID_FLIGHT_DATE
        - FLIGHT_PROVIDER_NOT_CONFIGURED
        - FLIGHT_PROVIDER_TIMEOUT
        - FLIGHT_PROVIDER_RATE_LIMITED
        - FLIGHT_PROVIDER_ERROR
        - INVALID_QR_TOKEN
        - QR_TOKEN_EXPIRED
        - QR_TOKEN_ALREADY_USED
        - QR_TOKEN_TYPE_MISMATCH
        - QR_TOKEN_BOOKING_MISMATCH
...
    ApiResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
...
    ErrorResponse:
      type: object
      required: [success, error_code, message]
      properties:
        success:
          type: boolean
          const: false
        error_code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    NotificationListItem:
      type: object
      properties:
        notificationId:
          type: integer
        notificationType:
          type: string
        title:
          type: string
        body:
          type: string
        payload:
          type: object
          additionalProperties: true
          description: Sanitized references only; no tokens or secrets
        read:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    PaginatedNotificationList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                page:
                  type: integer
                page_size:
                  type: integer
                total:
                  type: integer
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/NotificationListItem'
...